replace redis with nats
redis cluster and sentenal don't do lists / blpop very well
and run the risk of missing messages and duplicates
