+++ conf/index.js
@@ -7,7 +7,7 @@ module.exports = {
   , port: 3455
   }
 , PORT: 3000
-, redis: {
-    hosts:'127.0.0.1:6379'
+, nats: {
+    hosts:'127.0.0.1:4222'
   }
 }
