+++ lib/timer.js
@@ -7,11 +7,11 @@
  * @since 1.0.0
  * @requires debug
  * @requires skyring/lib/transports
- * @requires skyring/lib/redis
+ * @requires skyring/lib/nats
  * @requires skyring/lib/json
  */
 const transports = require('./transports')
-    , redis      = require('./redis')
+    , nats      = require('./nats')
     , json       = require('./json')
     , debug      = require('debug')('skyring:timer')
     , cache      = new Map()
@@ -127,32 +127,29 @@ exports.update = function update(id, body, cb){
  **/
 exports.shutdown = function shutdown(cb) {
   BAIL = true;
-  redis.disconnect(() => {
-    const exit = false;
-    const size = cache.size;
-    
-    if( !size ) return cb();
+  const exit = false;
+  const size = cache.size;
 
-    const client = redis.client;
-    let count = 0;
-    const run = (obj) => {
-      clearTimeout(record.timer);
+  if( !size ) return cb();
 
-      debug('pushing', record.id)
-      const data = Object.assign({}, record.payload, {
-        id: record.id
-      , created: record.created
-      });
+  const client = nats.client;
+  let count = 0;
+  const run = (obj) => {
+    clearTimeout(record.timer);
+    debug('pushing', record.id)
+    const data = Object.assign({}, record.payload, {
+      id: record.id
+    , created: record.created
+    });
 
-      client.lpush('skyring', JSON.stringify( data ));
-      if (++count === size) return client.quit(cb)
-    }
+    nats.client.publish('skyring', JSON.stringify(data));
+    if (++count === size) return setImmediate(nats.quit, cb)
+}
 
-    for(var record of cache.values()) {
-      run(record);
-    }
-    cache.clear();
-  })
+for(var record of cache.values()) {
+  run(record);
+}
+cache.clear();
 }
 
 /**
@@ -162,16 +159,12 @@ exports.shutdown = function shutdown(cb) {
  * @param {Nodeback} callback Node style callback to execute when the function has finished execution
  **/
 exports.watch = function( key, cb ){
-  (function pull(){
+  if(BAIL) return
+  debug('nats subscribe: skyring')
+  nats.client.subscribe('skyring', {queue: 'skyring'}, (data) => {
     if(BAIL) return
-    redis.client.blpop(key, 5, (err, replies) => {
-      if(BAIL) return
-      if (!err && !replies) return setImmediate(pull)
-      if(err) return setImmediate(cb, err);
-      const value = json.parse(replies[1])
-      setImmediate(pull);
-      cb(value.error, value.value);
-    });
-  }())
+    const value = json.parse(data)
+    cb(value.error, value.value);
+  });
 }
